added utils folder for calculation methods, moved poseDetector to utils
